Added LogoutTest + refactoring
